+++ src/Mod/Draft/DraftTools.py
@@ -40,7 +40,7 @@ __url__ = "http://www.freecadweb.org"
 
 import sys, os, FreeCAD, FreeCADGui, WorkingPlane, math, re, Draft, Draft_rc, DraftVecUtils
 from FreeCAD import Vector
-from DraftGui import todo,QtCore,QtGui
+from DraftGui import todo, QtCore, QtGui, translate, utf8_decode
 from DraftSnap import *
 from DraftTrackers import *
 from pivy import coin
@@ -72,51 +72,6 @@ MODALT = MODS[Draft.getParam("modalt",2)]
 #---------------------------------------------------------------------------
 # General functions
 #---------------------------------------------------------------------------
-
-
-try:
-    _encoding = QtGui.QApplication.UnicodeUTF8
-    def translate(context, text, utf8_decode=False):
-        """convenience function for Qt translator
-            context: str
-                context is typically a class name (e.g., "MyDialog")
-            text: str
-                text which gets translated
-            utf8_decode: bool [False]
-                if set to true utf8 encoded unicode will be returned. This option does not have influence
-                on python3 as for python3 we are returning utf-8 encoded unicode by default!
-        """
-        if sys.version_info.major >= 3 or utf8_decode:
-            return QtGui.QApplication.translate(context, text, None, _encoding)
-        else:
-            return QtGui.QApplication.translate(context, text, None, _encoding).encode("utf8")
-
-except AttributeError:
-    def translate(context, text, utf8_decode=False):
-        """convenience function for Qt translator
-            context: str
-                context is typically a class name (e.g., "MyDialog")
-            text: str
-                text which gets translated
-            utf8_decode: bool [False]
-                if set to true utf8 encoded unicode will be returned. This option does not have influence
-                on python3 as for python3 we are returning utf-8 encoded unicode by default!
-        """
-        if sys.version >= 3 or utf8_decode:
-            return QtGui.QApplication.translate(context, text, None)
-        else:
-            return QtGui.QApplication.translate(context, text, None).encode("utf8")
-
-def utf8_decode(text):
-    """py2: str     -> unicode
-            unicode -> unicode
-       py3: str     -> str
-            bytes   -> str
-    """
-    try:
-        return text.decode("utf-8")
-    except AttributeError:
-        return text
 
 def msg(text=None,mode=None):
     "prints the given message on the FreeCAD status bar"
